requirements.txt
selenium==4.16.0
PyYAML
cryptography